requirements.txt
cryptography~=44.0.0
pytest~=8.3.4
pymongo~=4.10.1
setuptools~=75.6.0
python-dotenv~=1.0.1